Move `Card` and `PageTemplate` into `src/site/components`
